+++ packages/kernel/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.3] - 2023-02-17
+### Changed
+- Use Croquet OS version 1.1.0-16.
+
 ## [1.3.2] - 2022-11-28
 ### Changed
 - Use Croquet OS version 1.1.0-14.